DRAFT luci-interpreter KernelBuilderHelper
Extract helper functions from KernelBuilder.
